[flaky-tests] fix testSend and testRecv in ProcessGroupGlooTest
Pull Request resolved: https://github.com/pytorch/pytorch/pull/32134

These tests weren't written in the most correct way and were often
flaky. It was tricky to identify these tests as flaky until we moved this file
to use gtest.

The gist of the issue is that the test previously would not coordinate sends
and recvs properly. For example, we created a single thread to test an
abortRecv and a successful recv. A separate sender thread was used to send 2
messages. What could go wrong here is that the first send could successfully
complete, resulting in the receiving end processing the message before it gets
the abort signal. In this case we would have an error in the test.
ghstack-source-id: 96675370

Differential Revision: [D19379395](https://our.internmc.facebook.com/intern/diff/D19379395/)